+++ HealthPlanet/src/main/java/com/dongduk/HealthPlanet/dao/ConnectionManager.java
@@ -1,4 +1,4 @@
-package com.dongduk.HealthPlanet.model.dao;
+package com.dongduk.HealthPlanet.dao;
 
 import java.io.IOException;
 import java.io.InputStream;
