+++ HealthPlanet/src/main/java/com/dongduk/HealthPlanet/domain/Wish.java
@@ -1,4 +1,4 @@
-package com.dongduk.HealthPlanet.model;
+package com.dongduk.HealthPlanet.domain;
 
 public class Wish {
     private int id;
